Refactored packages and added channels in README
