+++ bundles/org.openhab.binding.solax/src/main/java/org/openhab/binding/solax/internal/model/InverterType.java
@@ -18,11 +18,11 @@ import java.util.stream.Stream;
 
 import org.eclipse.jdt.annotation.NonNullByDefault;
 import org.eclipse.jdt.annotation.Nullable;
-import org.openhab.binding.solax.internal.model.local.parsers.RawDataParser;
-import org.openhab.binding.solax.internal.model.local.parsers.X1BoostAirMiniDataParser;
-import org.openhab.binding.solax.internal.model.local.parsers.X1HybridG4DataParser;
-import org.openhab.binding.solax.internal.model.local.parsers.X3HybridG4DataParser;
-import org.openhab.binding.solax.internal.model.local.parsers.X3MicOrProG2DataParser;
+import org.openhab.binding.solax.internal.model.local.RawDataParser;
+import org.openhab.binding.solax.internal.model.local.inverters.X1BoostAirMiniDataParser;
+import org.openhab.binding.solax.internal.model.local.inverters.X1HybridG4DataParser;
+import org.openhab.binding.solax.internal.model.local.inverters.X3HybridG4DataParser;
+import org.openhab.binding.solax.internal.model.local.inverters.X3MicOrProG2DataParser;
 
 /**
  * The {@link InverterType} class is enum representing the different inverter types with a simple logic to convert from
